Refactor mobile app components and clean up unused files
