+++ mobile/src/app/(checkout)/order-review.tsx
@@ -3,7 +3,7 @@ import { View, Text, StyleSheet, ScrollView } from 'react-native'
 import { useCheckout } from '../../context/CheckoutContext'
 import { router, Stack } from 'expo-router'
 import { Ionicons } from '@expo/vector-icons'
-import { TouchableOpacity } from '@/src/components/ui/TouchableOpacity'
+import { TouchableOpacity } from '@/src/components/Themed'
 import { useCustomTheme } from '@/src/context/ThemeContext'
 import { Theme } from '@/src/types/theme'
 import StatusCheckout from '@/src/components/front/StatusCheckout'
