+++ mobile/src/app/(checkout)/payment.tsx
@@ -2,14 +2,11 @@ import React, { useState } from 'react'
 import { ScrollView, StyleSheet, Alert } from 'react-native'
 import { Ionicons } from '@expo/vector-icons'
 import { Stack, useRouter } from 'expo-router'
-import { TouchableOpacity } from '@/src/components/ui/TouchableOpacity'
-
-import { Text, View } from '@/src/components/Themed'
+import { Text, View, TouchableOpacity } from '@/src/components/Themed'
 import { useCustomTheme } from '@/src/context/ThemeContext'
 import { Theme } from '@/src/types/theme'
 import StatusCheckout from '@/src/components/front/StatusCheckout'
-import TextInput from '@/src/components/ui/TextInput'
-import FileInput from '@/src/components/ui/FileInput'
+import { TextInput, FileInput } from '@/src/components/Themed'
 import { useStateStore } from '@/src/store/store'
 import { useShallow } from 'zustand/react/shallow'
