+++ mobile/src/app/(main)/(settings)/index.tsx
@@ -7,7 +7,7 @@ import { FontSize } from '@/src/constants/FontSize'
 import { useRouter } from 'expo-router'
 import Colors from '@/src/constants/Theme'
 import { useState } from 'react'
-import ModalMessage from '@/src/components/ui/ModalMessage'
+import ModalMessage from '@/src/components/modal/ModalMessage'
 import Constants from 'expo-constants'
 import { useCustomTheme } from '@/src/context/ThemeContext'
 import { Theme } from '@/src/types/theme'
