+++ mobile/src/app/(main)/index.tsx
@@ -9,8 +9,7 @@ import { FlatList, ScrollView, StyleSheet } from 'react-native'
 import { Separador } from '@/src/components/front/Separador'
 import ServicesCard from '@/src/components/front/ServicesCard'
 import CompletedServiceCard from '../../components/front/CompletedServiceCard'
-import { View } from '@/src/components/Themed'
-import { Text } from '@/src/components/ui/Text'
+import { View,Text } from '@/src/components/Themed'
 const users = [
 	{
 		id: 1,
