+++ mobile/src/app/(services)/completed-service-details.tsx
@@ -11,13 +11,11 @@ import {
 	Platform,
 } from 'react-native'
 import FastImage from 'react-native-fast-image'
-import { fontFamily } from '@/src/constants/FontFamily'
 import { Ionicons, MaterialIcons, FontAwesome } from '@expo/vector-icons'
-import Colors from '@/src/constants/Theme'
 import { FontSize } from '@/src/constants/FontSize'
 import { LinearGradient } from 'react-native-linear-gradient'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
-import ModalMessage from '@/src/components/ui/ModalMessage'
+import ModalMessage from '@/src/components/modal/ModalMessage'
 import { Separador } from '@/src/components/front/Separador'
 import Star from '@/src/components/front/Star'
 import { Theme } from '@/src/types/theme'
