+++ mobile/src/app/(user)/edit-profile.tsx
@@ -18,11 +18,11 @@ import { Stack, useRouter } from 'expo-router'
 import { useState, useRef, useEffect } from 'react'
 import NetInfo from '@react-native-community/netinfo'
 import * as ImagePicker from 'expo-image-picker'
-import TextInputUI from '@/src/components/ui/TextInput'
-import ModalMessage from '@/src/components/ui/ModalMessage'
+import { TextInput } from '@/src/components/Themed'
+import ModalMessage from '@/src/components/modal/ModalMessage'
 import { useCustomTheme } from '@/src/context/ThemeContext'
 import { Theme } from '@/src/types/theme'
-import { TouchableOpacity } from '@/src/components/ui/TouchableOpacity'
+import { TouchableOpacity } from '@/src/components/Themed'
 export default function EditProfileScreen() {
 	const { user } = useUser()
 	const { session } = useSession()
@@ -253,7 +253,7 @@ export default function EditProfileScreen() {
 
 								<View style={styles.formContainer}>
 									<View style={styles.inputGroup}>
-										<TextInputUI
+										<TextInput
 											type="text"
 											label="Nome"
 											icon="person-outline"
@@ -264,7 +264,7 @@ export default function EditProfileScreen() {
 									</View>
 
 									<View style={styles.inputGroup}>
-										<TextInputUI
+										<TextInput
 											type="text"
 											icon="person-outline"
 											label="Sobrenome"
