+++ mobile/src/components/modal/ModalSSO.tsx
@@ -6,15 +6,13 @@ import {
 } from 'react-native'
 import { Text, View } from '../Themed'
 import { TextInput } from 'react-native'
-import { FontSize } from '@/src/constants/FontSize'
 import { Ionicons } from '@expo/vector-icons'
-import { Pressable } from 'react-native'
 import { useEffect, useRef, useState } from 'react'
 import { ClerkAPIError } from '@clerk/types'
 import { router } from 'expo-router'
 import * as Haptics from 'expo-haptics'
 import { isClerkAPIResponseError, useSignUp } from '@clerk/clerk-expo'
-import { TouchableOpacity } from './TouchableOpacity'
+import { TouchableOpacity } from '../Themed'
 import { maskEmail } from '@/src/utils/maskEmail'
 import { useCustomTheme } from '@/src/context/ThemeContext'
 import { Theme } from '@/src/types/theme'
