--- mobile/src/components/ui/BodyScrollView.tsx
@@ -1,15 +1,0 @@
-import { forwardRef } from 'react'
-import { ScrollView, ScrollViewProps } from 'react-native'
-
-export const BodyScrollView = forwardRef<any, ScrollViewProps>((props, ref) => {
-	return (
-		<ScrollView
-			automaticallyAdjustsScrollIndicatorInsets
-			contentInsetAdjustmentBehavior="automatic"
-			contentInset={{ bottom: 0 }}
-			scrollIndicatorInsets={{ bottom: 0 }}
-			{...props}
-			ref={ref}
-		/>
-	)
-})
